networks/urls.py
from .views import ElectronicsNetworkViewSet
from rest_framework.routers import SimpleRouter
from django.urls import path, include
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from rest_framework.permissions import AllowAny

app_name = 'networks'

router = SimpleRouter()
router.register(r'networks', ElectronicsNetworkViewSet)

urlpatterns = [
    path(
        "login/",
        TokenObtainPairView.as_view(permission_classes=(AllowAny,)),
        name="login",
    ),
    path(
        "token/refresh/",
        TokenRefreshView.as_view(permission_classes=(AllowAny,)),
        name="token_refresh",
    ),
] + router.urls